Extract public single share UI helper
