+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{M as Ee,a as vt,b as D,A as d,c as y,o as W,n as Ct,d as Et}from"./oc-compat-CCOQZroc.chunk.mjs";import{p as T,a as xt}from"./urls-COV-UP-R.chunk.mjs";(function(){const K="etherpad_nextcloud_open",h=new Map;let V=!1,G=null,J=null,Z=0,Q="",A=null,F=null,P=0,X=0,q=!1,b=!1,Y=!1,ee=0;const U="data-epnc-sidebar-sync-panel",j="data-epnc-sidebar-sync-mount",te=8e3,xe=6e4,ne=15e3,Ae=8e3,C="data-epnc-single-file-share",ie=d+"_public_pad",re="icon-filetype-etherpad-nextcloud-pad",oe='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',ae=98,Ne=120,Le=500,se=()=>{const e=Ct();return e!==null&&e>=33},E=(e,n)=>{const i=!e||e==="/"?"":String(e),r=String(n||"").replace(/^\/+/,"").replace(/\s+\.pad$/i,".pad");return i===""?"/"+r:i+"/"+r},Se=e=>y("/apps/"+d+"/?file="+encodeURIComponent(e)),x=(e,n)=>{const i=y("/apps/files/files/"+encodeURIComponent(String(e))),r=new URLSearchParams;return r.set("dir",fe(n)),r.set("editing","false"),r.set("openfile","true"),i+"?"+r.toString()},N=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),f=()=>(window.location.pathname||"").includes("/apps/files"),le=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=n.pathname||"",r=i.match(/\/apps\/files\/files\/(\d+)\/?$/);if(!r){const a=i.match(/\/f\/(\d+)\/?$/);if(!a){const l=Number(n.searchParams.get("fileid")||"");return Number.isFinite(l)&&l>0?l:null}const s=parseInt(a[1],10);return Number.isFinite(s)&&s>0?s:null}const o=parseInt(r[1],10);return Number.isFinite(o)&&o>0?o:null},L=()=>{const e=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!e)return null;const n=parseInt(e[1],10);return Number.isFinite(n)&&n>0?n:null},O=(e,n)=>{const i=y("/apps/"+d+"/public/"+encodeURIComponent(e));return n?i+"?file="+encodeURIComponent(n):i},ke=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);if(!i)return null;const r=(n.searchParams.get("files")||"").trim();if(!S(r))return null;const o=(n.searchParams.get("path")||"/").trim()||"/";return{token:i[1],path:E(o,r)}},ce=e=>{if(!e)return null;const n='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',i=document.querySelector(n);return i instanceof HTMLAnchorElement?i:null},Ie=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const n=String(e.value||"").trim();if(n==="")return"";try{const i=window.atob(n);if(i.startsWith('"')||i.startsWith("{")||i.startsWith("[")){const r=JSON.parse(i);return typeof r=="string"?r:""}return i}catch{return""}},de=e=>{if(!e)return!1;const n=ce(e);if(!n)return!1;let i;try{i=new URL(n.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(i.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==e?!1:!i.searchParams.has("files")},Me=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const n=T();if(!n)return e.removeAttribute(C),{retry:!1};const i=Ie();return i==="public-file-share"?(e.setAttribute(C,"1"),{retry:!1}):i==="public-share"?(e.removeAttribute(C),{retry:!1}):ce(n)?de(n)?(e.setAttribute(C,"1"),{retry:!1}):(e.removeAttribute(C),{retry:!1}):(e.removeAttribute(C),{retry:!0})},Te=()=>{const e=++X,n=i=>{e===X&&Me().retry&&i<30&&window.setTimeout(()=>n(i+1),180)};n(0)},S=e=>typeof e=="string"&&e.toLowerCase().endsWith(".pad"),k=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),Fe=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const n=Et(),i=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of i){const o=Number(r);if(Number.isFinite(o))return(o&n)===n}return null},Pe=(...e)=>{for(const n of e){const i=Fe(n);if(i!==null)return i}return!0},qe=(...e)=>Pe(...e),ue=()=>{at()},pe=()=>({id:ie,displayName:t(d,"Public pad"),...se()?{iconSvgInline:oe}:{iconClass:re},order:ae,enabled:qe,handler:ue}),Oe=()=>({id:ie,displayName:t(d,"Public pad"),templateName:t(d,"Public pad"),iconClass:re,...se()?{iconSvgInline:oe}:{},fileType:"file",order:ae,actionHandler:ue}),_=async e=>{const n=String(e);if(h.has(n))return h.get(n);const i=y("/apps/"+d+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(n),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve failed.");return a}).catch(o=>{throw h.delete(n),o});return h.set(n,r),r},_e=async e=>{const n="path:"+String(e);if(h.has(n))return h.get(n);const i=y("/apps/"+d+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve by path failed.");return a}).catch(o=>{throw h.delete(n),o});return h.set(n,r),r},He=e=>{if(!e||typeof e!="string")return"/";const n=e.lastIndexOf("/");return n<=0?"/":e.slice(0,n)||"/"},fe=e=>{const n=He(e);if(n!=="/")return n;const i=H();return i===""?"/":i},H=()=>{const e=(new URLSearchParams(window.location.search||"").get("dir")||"/").trim();return e===""?"/":e},Re=()=>{const e=document.documentElement,n=document.body,i=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(i))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},me=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},Ue=()=>{const e=L();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),i=Number(n.get("fileid")||"");return Number.isFinite(i)&&i>0?i:null},R=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},je=e=>{if(!e)return null;const n=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const i of n){const r=R(i);if(r!==null)return r}if(typeof e.get=="function")for(const i of["fileid","fileId","id"])try{const r=R(e.get(i));if(r!==null)return r}catch{}return null},ze=e=>{if(!(e instanceof Element))return null;const n=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of n){const o=R(r);if(o!==null)return o}const i=e.matches("a[href]")?e:e.querySelector("a[href]");return i instanceof HTMLAnchorElement?le(i.getAttribute("href")):null},Be=e=>{if(!e||typeof e!="object")return null;const n=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of n){const o=je(r);if(o!==null)return o}const i=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of i){const o=ze(r);if(o!==null)return o}return null},ye=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const i=document.querySelector(n);if(i instanceof HTMLElement)return i}return null},$e=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+j+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute(j,"1"));const i=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(i&&i.parentNode instanceof HTMLElement){const r=i.parentNode;return(n.parentNode!==r||n.nextSibling!==i)&&r.insertBefore(n,i),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},we=e=>{const n=Ue();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const i=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const o of i){const a=e.querySelector(o);if(!(a instanceof HTMLElement))continue;const s=[a.getAttribute("data-fileid"),a.getAttribute("data-file-id"),a.getAttribute("data-node-id"),a.getAttribute("data-id"),a.dataset&&(a.dataset.fileid||a.dataset.fileId)];for(const l of s){const c=R(l);if(c!==null)return c}}const r=e.querySelectorAll("a[href]");for(const o of r){if(!(o instanceof HTMLAnchorElement))continue;const a=le(o.getAttribute("href"));if(a!==null)return a}return null},De=e=>e==="synced"?"etherpad-icon-color":Re()?"etherpad-icon-white":"etherpad-icon-black",We=e=>e==="synced"?t(d,"Synchronized"):e==="syncing"?t(d,"Saving..."):e==="error"?t(d,"Sync failed"):e==="unavailable"?t(d,"Sync status unavailable"):e==="loading"?t(d,"Checking sync status..."):t(d,"Sync pending"),v=()=>{const e=document.querySelector("["+U+'="1"]');e&&e.parentNode&&e.parentNode.removeChild(e);const n=document.querySelector("["+j+'="1"]');n&&n.parentNode&&n.parentNode.removeChild(n),I()},Ke=(e,n)=>{if(!(e instanceof HTMLElement))return;const i=String(n||"").trim();e.setAttribute("data-public-open-url",i);const r=e.querySelector("[data-epnc-sidebar-open-public-button]");if(!(r instanceof HTMLButtonElement))return;const o=i!=="";r.style.display=o?"inline-flex":"none",r.disabled=!o},g=(e,n,i,r,o=!0)=>{if(!(e instanceof HTMLElement))return;const a=e.querySelector("[data-epnc-sidebar-sync-icon]"),s=e.querySelector("[data-epnc-sidebar-sync-label]"),l=e.querySelector("[data-epnc-sidebar-sync-button]");if(e.setAttribute("data-file-id",String(i)),e.setAttribute("data-sync-state",n),e.setAttribute("data-sync-in-flight",r?"1":"0"),s instanceof HTMLElement&&(s.textContent=We(n)),a instanceof HTMLElement){const c=De(n),u=D(d,c);u!==""&&(a.style.backgroundImage=`url(${u})`)}l instanceof HTMLButtonElement&&(l.disabled=r||!o||i<=0,l.setAttribute("aria-busy",r?"true":"false"))},Ve=(e,n)=>{if(!(n instanceof HTMLElement))return null;let i=n.querySelector("["+U+'="1"]');if(!(i instanceof HTMLElement)){i=document.createElement("section"),i.className="epnc-sidebar-sync-panel",i.setAttribute(U,"1"),i.setAttribute("data-file-id",String(e));const r=document.createElement("div");r.className="epnc-sidebar-sync-panel__header";const o=document.createElement("span");o.className="epnc-sidebar-sync-panel__icon",o.setAttribute("data-epnc-sidebar-sync-icon","1"),o.setAttribute("aria-hidden","true");const a=document.createElement("span");a.className="epnc-sidebar-sync-panel__status",a.setAttribute("data-epnc-sidebar-sync-label","1"),r.appendChild(o),r.appendChild(a);const s=document.createElement("button");s.type="button",s.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",s.setAttribute("data-epnc-sidebar-sync-button","1"),s.textContent=t(d,"Pad in Datei speichern"),s.addEventListener("click",async c=>{c.preventDefault(),c.stopPropagation();const u=Number(i.getAttribute("data-file-id")||"");if(!(!Number.isFinite(u)||u<=0)){g(i,"syncing",u,!0,!0);try{await Ge(u,!0);const m=await B(u),w=z(m&&m.status);g(i,w,u,!1,!0)}catch{g(i,"error",u,!1,!0)}}});const l=document.createElement("button");l.type="button",l.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",l.setAttribute("data-epnc-sidebar-open-public-button","1"),l.textContent=t(d,"Pad in neuem Tab öffnen"),l.style.display="none",l.addEventListener("click",c=>{c.preventDefault(),c.stopPropagation();const u=String(i.getAttribute("data-public-open-url")||"").trim();u!==""&&window.open(u,"_blank","noopener,noreferrer")}),i.appendChild(r),i.appendChild(s),i.appendChild(l)}return i.parentNode!==n&&n.appendChild(i),i},z=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",B=async e=>{const n=y("/apps/"+d+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e))),i=await fetch(n,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}),r=await i.json().catch(()=>({}));if(!i.ok)throw new Error(r&&r.message||"Sync status check failed.");return r},Ge=async(e,n)=>{let i=y("/apps/"+d+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));i+="?force=1";const r=await fetch(i,{method:"POST",headers:{Accept:"application/json",requesttoken:W()},credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Pad sync failed.");return o},I=()=>{F!==null&&(window.clearTimeout(F),F=null)},Je=(e,n)=>e==="synced"?Math.min(xe,n+Ae):e==="pending"?te:ne,he=(e,n)=>{if(I(),!(e instanceof HTMLElement)||!Number.isFinite(n)||n<=0)return;let i=te;const r=a=>{F=window.setTimeout(()=>{o()},Math.max(1e3,a))},o=async()=>{if(!(e instanceof HTMLElement)||!document.body.contains(e)){I();return}const a=ye();if(!(a instanceof HTMLElement)||!me(a)||!me(e)){I();return}if(we(a)!==n){I();return}if(document.visibilityState!=="visible"||e.getAttribute("data-sync-in-flight")==="1"){r(i);return}try{const s=await B(n),l=z(s&&s.status);g(e,l,n,!1,!0),i=Je(l,i)}catch{g(e,"error",n,!1,!0),i=ne}r(i)};r(i)},Ze=()=>{A!==null&&(window.clearTimeout(A),A=null)},Qe=async()=>{const e=++P;if(!f()||T()){v();return}const n=ye();if(!(n instanceof HTMLElement)){v();return}const i=we(n);if(!i){v();return}const r=$e(n);if(!(r instanceof HTMLElement)){v();return}let o;try{o=await _(i)}catch{v();return}if(e!==P)return;if(!o||o.is_pad!==!0){v();return}const a=Ve(i,r);if(!(a instanceof HTMLElement)){v();return}const s=o&&typeof o.public_open_url=="string"?o.public_open_url.trim():"";Ke(a,s),g(a,"loading",i,!1,!0);try{const l=await B(i);if(e!==P)return;const c=z(l&&l.status);g(a,c,i,!1,!0),he(a,i)}catch{if(e!==P)return;g(a,"error",i,!1,!0),he(a,i)}},$=(e=0)=>{Ze(),A=window.setTimeout(()=>{A=null,Qe()},Math.max(0,e))},Xe=e=>{try{const n=new URL(e,window.location.origin),i=decodeURIComponent(n.pathname||"").match(/\/p\/([^/]+)$/);if(i&&i[1]){const r=i[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},Ye=async(e,n)=>{const i=y("/apps/"+d+"/api/v1/pads/from-url"),r=new URLSearchParams;r.set("file",e),r.set("padUrl",n);const o=await fetch(i,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:W()},credentials:"same-origin",body:r.toString()}),a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Could not import public pad URL.");return a},et=async e=>{const n=y("/apps/"+d+"/api/v1/pads"),i=new URLSearchParams;i.set("file",e),i.set("accessMode","public");const r=await fetch(n,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:W()},credentials:"same-origin",body:i.toString()}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Could not create public pad.");return o},be=e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(d,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",i.appendChild(r),i.appendChild(o),n.appendChild(i),document.body.appendChild(n),{overlay:n,dialog:i,closeButton:r}},tt=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=be(t(d,"Internal public pad")),o=document.createElement("label");o.textContent=t(d,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(d,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const s=document.createElement("p");s.style.color="var(--color-error, #c62828)",s.style.margin="0 0 12px 0",s.style.minHeight="20px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(d,"Create");const c=u=>{n.remove(),e(u)};r.addEventListener("click",()=>c(null)),n.addEventListener("click",u=>{u.target===n&&c(null)}),a.addEventListener("keydown",u=>{u.key==="Enter"&&(u.preventDefault(),l.click())}),l.addEventListener("click",()=>{const u=a.value.trim();if(u===""){s.textContent=t(d,"File name is required.");return}c(u)}),i.appendChild(o),i.appendChild(a),i.appendChild(s),i.appendChild(l),a.focus(),a.select()}),nt=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=be(t(d,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(d,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const s=document.createElement("label");s.textContent=t(d,"File name"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="text",l.value="Imported pad.pad",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const c=document.createElement("p");c.style.color="var(--color-error, #c62828)",c.style.margin="0 0 12px 0",c.style.minHeight="20px";const u=document.createElement("button");u.type="button",u.className="primary",u.textContent=t(d,"Create");const m=p=>{n.remove(),e(p)};r.addEventListener("click",()=>m(null)),n.addEventListener("click",p=>{p.target===n&&m(null)}),a.addEventListener("blur",()=>{const p=a.value.trim();p.startsWith("http")&&(l.value=Xe(p))});const w=()=>{const p=a.value.trim(),Ce=l.value.trim();if(p===""){c.textContent=t(d,"External pad URL is required.");return}if(Ce===""){c.textContent=t(d,"File name is required.");return}m({padUrl:p,name:Ce})};a.addEventListener("keydown",p=>{p.key==="Enter"&&(p.preventDefault(),w())}),l.addEventListener("keydown",p=>{p.key==="Enter"&&(p.preventDefault(),w())}),u.addEventListener("click",w),i.appendChild(o),i.appendChild(a),i.appendChild(s),i.appendChild(l),i.appendChild(c),i.appendChild(u),a.focus(),a.select()}),it=async()=>{const e=await tt();if(!e)return;const n=e.trim(),i=n.toLowerCase().endsWith(".pad")?n:n+".pad",r=E(H(),i);try{const o=await et(r),a=o&&typeof o.file=="string"?o.file:r,s=o&&Number.isFinite(Number(o.file_id))?Number(o.file_id):null;if(f()&&s!==null){window.location.assign(x(s,a));return}await M({path:a,fileId:s})}catch(o){const a=o instanceof Error?o.message:t(d,"Could not create public pad.");window.alert(a)}},rt=async()=>{const e=await nt();if(!e)return;const n=e.padUrl.trim(),i=e.name.trim(),r=H(),o=i.toLowerCase().endsWith(".pad")?i:i+".pad",a=E(r,o);try{const s=await Ye(a,n),l=s&&typeof s.file=="string"?s.file:a,c=s&&Number.isFinite(Number(s.file_id))?Number(s.file_id):null;if(f()&&c!==null){window.location.assign(x(c,l));return}await M({path:l,fileId:c})}catch(s){const l=s instanceof Error?s.message:t(d,"Could not import public pad URL.");window.alert(l)}},ot=()=>new Promise(e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(d,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(d,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(d,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const s=document.createElement("div");s.style.display="grid",s.style.gap="8px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(d,"Internal public pad");const c=document.createElement("button");c.type="button",c.textContent=t(d,"External public pad (URL)");const u=D(d,"etherpad-icon-color");u!==""&&(c.style.backgroundImage=`url(${u})`),c.style.backgroundRepeat="no-repeat",c.style.backgroundPosition="12px center",c.style.backgroundSize="16px 16px",c.style.paddingLeft="34px";const m=w=>{n.remove(),e(w)};o.addEventListener("click",()=>m(null)),l.addEventListener("click",()=>m("internal")),c.addEventListener("click",()=>m("external")),n.addEventListener("click",w=>{w.target===n&&m(null)}),s.appendChild(l),s.appendChild(c),i.appendChild(o),i.appendChild(r),i.appendChild(a),i.appendChild(s),n.appendChild(i),document.body.appendChild(n)}),at=async()=>{const e=await ot();if(e==="internal"){await it();return}e==="external"&&await rt()},st=(e,n)=>{const i=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!i||typeof i.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(i.goToRoute(null,{view:"files",fileid:String(e)},{dir:fe(n),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=x(e,n);window.location.assign(r)},180),!0)},M=async e=>{const n=String(e.fileId??"")+"|"+String(e.path??""),i=Date.now();if(J===n&&i-Z<800)return;J=n,Z=i;const r=T(),o=r!==null&&r!=="",a=()=>{if(o){window.location.assign(O(r,e.path||""));return}if(f()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const u=e.path||"/";window.location.assign(x(Number(e.fileId),u));return}const c=f()?L():null;if((e.fileId===null||e.fileId===void 0)&&c){const u=e.path||"/";window.location.assign(x(c,u));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const u=e.path||"/";window.location.assign(x(Number(e.fileId),u));return}e.path&&window.location.assign(Se(e.path))};if(!N()){a();return}let s=e.path||"",l=e.fileId??null;if(!s&&e.fileId!==null&&e.fileId!==void 0)try{const c=await _(e.fileId);s=c&&typeof c.path=="string"?c.path:"",l=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):l}catch{s=""}if(!s){a();return}if((!l||!Number.isFinite(Number(l)))&&s&&!o)try{const c=await _e(s);l=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):l}catch{}if((!l||!Number.isFinite(Number(l)))&&f()){const c=L();c&&(l=c)}if(f()){if(l&&Number.isFinite(Number(l))&&st(Number(l),s))return;a();return}try{window.OCA.Viewer.open({path:s})}catch{a()}},lt=(e,n)=>{if(!(e instanceof HTMLAnchorElement)||!n)return null;const i=e.getAttribute("href")||"",r=ke(i);if(r&&r.token===n&&S(r.path))return r.path;const o=xt(i);return S(o)?o:null},ct=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const e=async n=>{if(!n||n.defaultPrevented)return;const i=T();if(!i||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const r=n.target.closest("a[href]");if(!(r instanceof HTMLAnchorElement))return;const o=lt(r,i);if(o){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),N()){await M({path:o,fileId:null});return}window.location.assign(O(i,o))}};document.addEventListener("click",e,!0)},dt=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(Ee,K,vt(),D(d,"etherpad-icon-color"),(e,n)=>{const i=n&&n.dir||H(),r=E(i,e),o=Be(n);M({path:r,fileId:o})},t(d,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(Ee,K))},ut=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const i of e)if(!(!i||typeof i!="object")){if(typeof i.addNewFileMenuEntry=="function")return{register:i.addNewFileMenuEntry.bind(i),unregister:typeof i.removeNewFileMenuEntry=="function"?i.removeNewFileMenuEntry.bind(i):null};if(typeof i.getNewFileMenu=="function"){let r=null;try{r=i.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const n=window._nc_newfilemenu;return n&&typeof n=="object"&&typeof n.registerEntry=="function"?{register:n.registerEntry.bind(n),unregister:typeof n.unregisterEntry=="function"?n.unregisterEntry.bind(n):null}:null},pt=async()=>{const e=ut();if(!e)return!1;const n=pe();try{return e.register(n),q=!0,!0}catch(i){return k(i)?(q=!0,!0):!1}},ft=()=>{if(b)return!0;const e=Oe(),n=()=>{const i=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const r of i)if(!(!r||typeof r!="object")){if(typeof r.addMenuEntry=="function")try{return r.addMenuEntry(e),!0}catch(o){if(k(o))return!0}if(typeof r.registerEntry=="function")try{return r.registerEntry(pe()),!0}catch(o){if(k(o))return!0}}return!1};try{return n()?(b=!0,!0):(!Y&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(i){if(!(!i||typeof i.addMenuEntry!="function"))try{i.addMenuEntry(e),b=!0}catch(r){k(r)&&(b=!0)}}}),Y=!0),b)}catch(i){return k(i)?(b=!0,!0):!1}},mt=()=>{if(!f()||q||b)return;const e=++ee,n=async i=>{if(!(e!==ee||q||b||await pt()||ft())&&i<Ne){window.setTimeout(()=>{n(i+1)},Le);return}};n(0)},yt=()=>{if(G!==null||!f())return;const e=new MutationObserver(n=>{for(const i of n)for(const r of i.addedNodes)if(r instanceof Element&&(r.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||r.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){$(60);return}});e.observe(document.body,{childList:!0,subtree:!0}),G=e},wt=async()=>{if(N())return;const e=L();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("epNoRedirect")==="1"||n.get("details")==="1"||n.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+d+"/"))try{const i=await _(e);i&&i.is_pad&&i.viewer_url&&window.location.assign(i.viewer_url)}catch{}},ht=async()=>{if(N()||!f())return;const e=L();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("openfile")==="true"||n.get("opendetails")==="true"||n.get("details")==="1"))try{const i=await _(e);if(!i||!i.is_pad)return;const r=n.get("dir")||"/",o=y("/apps/files/files")+"?dir="+encodeURIComponent(r);window.location.href!==o&&window.location.assign(o)}catch{}},bt=()=>{const e=T();if(!e)return;const n=new URLSearchParams(window.location.search||""),i=n.get("files")||"";if(N()){if(!S(i))return;const a=n.get("path")||"/",s=E(a,i);M({path:s,fileId:null});return}if(!S(i)){if((window.location.pathname||"").includes("/apps/"+d+"/public/"))return;de(e)&&window.location.assign(O(e,""));return}const r=n.get("path")||"/",o=E(r,i);window.location.assign(O(e,o))},ge=()=>{const e=(window.location.pathname||"")+"?"+(window.location.search||"");e!==Q&&(Q=e,f()&&mt(),Te(),ht(),wt(),bt(),$(0))},gt=()=>{const e="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[e]===!0)return;window.OCA&&(window.OCA[e]=!0);const n=()=>window.setTimeout(ge,0);window.addEventListener("popstate",n),window.addEventListener("hashchange",n);const i=r=>{const o=window.history&&window.history[r];typeof o=="function"&&(window.history[r]=function(...a){const s=o.apply(this,a);return n(),s})};i("pushState"),i("replaceState")},ve=()=>{V||(V=!0,gt(),ge(),dt(),ct(),yt(),$(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",ve,{once:!0}):ve()})();
+import{M as ge,a as ht,b as V,A as d,c as b,o as G,n as bt,d as gt}from"./oc-compat-CCOQZroc.chunk.mjs";import{p as q,a as vt}from"./urls-COV-UP-R.chunk.mjs";const L="data-epnc-single-file-share";let ve=0;const Ce=p=>{if(!p)return null;const f='a[href*="/s/'+p+'/download"],a[href*="/index.php/s/'+p+'/download"]',m=document.querySelector(f);return m instanceof HTMLAnchorElement?m:null},Ct=()=>{const p=document.getElementById("initial-state-files_sharing-view");if(!(p instanceof HTMLInputElement))return"";const f=String(p.value||"").trim();if(f==="")return"";try{const m=window.atob(f);if(m.startsWith('"')||m.startsWith("{")||m.startsWith("[")){const v=JSON.parse(m);return typeof v=="string"?v:""}return m}catch{return""}},Ee=p=>{if(!p)return!1;const f=Ce(p);if(!f)return!1;let m;try{m=new URL(f.getAttribute("href")||"",window.location.origin)}catch{return!1}const v=(m.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!v||v[1]!==p?!1:!m.searchParams.has("files")},Et=()=>{const p=document.body;if(!(p instanceof HTMLElement))return{retry:!0};const f=q();if(!f)return p.removeAttribute(L),{retry:!1};const m=Ct();return m==="public-file-share"?(p.setAttribute(L,"1"),{retry:!1}):m==="public-share"?(p.removeAttribute(L),{retry:!1}):Ce(f)?Ee(f)?(p.setAttribute(L,"1"),{retry:!1}):(p.removeAttribute(L),{retry:!1}):(p.removeAttribute(L),{retry:!0})},xt=()=>{const p=++ve,f=m=>{p===ve&&Et().retry&&m<30&&window.setTimeout(()=>f(m+1),180)};f(0)};(function(){const p="etherpad_nextcloud_open",f=new Map;let m=!1,v=null,J=null,Z=0,Q="",S=null,O=null,_=0,H=!1,C=!1,X=!1,Y=0;const B="data-epnc-sidebar-sync-panel",$="data-epnc-sidebar-sync-mount",ee=8e3,xe=6e4,te=15e3,Ae=8e3,ne=d+"_public_pad",ie="icon-filetype-etherpad-nextcloud-pad",re='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',oe=98,Ne=120,Le=500,ae=()=>{const e=bt();return e!==null&&e>=33},A=(e,n)=>{const i=!e||e==="/"?"":String(e),r=String(n||"").replace(/^\/+/,"").replace(/\s+\.pad$/i,".pad");return i===""?"/"+r:i+"/"+r},Se=e=>b("/apps/"+d+"/?file="+encodeURIComponent(e)),N=(e,n)=>{const i=b("/apps/files/files/"+encodeURIComponent(String(e))),r=new URLSearchParams;return r.set("dir",de(n)),r.set("editing","false"),r.set("openfile","true"),i+"?"+r.toString()},k=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),w=()=>(window.location.pathname||"").includes("/apps/files"),se=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=n.pathname||"",r=i.match(/\/apps\/files\/files\/(\d+)\/?$/);if(!r){const a=i.match(/\/f\/(\d+)\/?$/);if(!a){const l=Number(n.searchParams.get("fileid")||"");return Number.isFinite(l)&&l>0?l:null}const s=parseInt(a[1],10);return Number.isFinite(s)&&s>0?s:null}const o=parseInt(r[1],10);return Number.isFinite(o)&&o>0?o:null},I=()=>{const e=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!e)return null;const n=parseInt(e[1],10);return Number.isFinite(n)&&n>0?n:null},R=(e,n)=>{const i=b("/apps/"+d+"/public/"+encodeURIComponent(e));return n?i+"?file="+encodeURIComponent(n):i},ke=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);if(!i)return null;const r=(n.searchParams.get("files")||"").trim();if(!M(r))return null;const o=(n.searchParams.get("path")||"/").trim()||"/";return{token:i[1],path:A(o,r)}},M=e=>typeof e=="string"&&e.toLowerCase().endsWith(".pad"),T=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),Ie=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const n=gt(),i=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of i){const o=Number(r);if(Number.isFinite(o))return(o&n)===n}return null},Me=(...e)=>{for(const n of e){const i=Ie(n);if(i!==null)return i}return!0},Te=(...e)=>Me(...e),le=()=>{it()},ce=()=>({id:ne,displayName:t(d,"Public pad"),...ae()?{iconSvgInline:re}:{iconClass:ie},order:oe,enabled:Te,handler:le}),Fe=()=>({id:ne,displayName:t(d,"Public pad"),templateName:t(d,"Public pad"),iconClass:ie,...ae()?{iconSvgInline:re}:{},fileType:"file",order:oe,actionHandler:le}),U=async e=>{const n=String(e);if(f.has(n))return f.get(n);const i=b("/apps/"+d+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(n),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve failed.");return a}).catch(o=>{throw f.delete(n),o});return f.set(n,r),r},Pe=async e=>{const n="path:"+String(e);if(f.has(n))return f.get(n);const i=b("/apps/"+d+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve by path failed.");return a}).catch(o=>{throw f.delete(n),o});return f.set(n,r),r},qe=e=>{if(!e||typeof e!="string")return"/";const n=e.lastIndexOf("/");return n<=0?"/":e.slice(0,n)||"/"},de=e=>{const n=qe(e);if(n!=="/")return n;const i=j();return i===""?"/":i},j=()=>{const e=(new URLSearchParams(window.location.search||"").get("dir")||"/").trim();return e===""?"/":e},Oe=()=>{const e=document.documentElement,n=document.body,i=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(i))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},ue=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},_e=()=>{const e=I();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),i=Number(n.get("fileid")||"");return Number.isFinite(i)&&i>0?i:null},z=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},He=e=>{if(!e)return null;const n=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const i of n){const r=z(i);if(r!==null)return r}if(typeof e.get=="function")for(const i of["fileid","fileId","id"])try{const r=z(e.get(i));if(r!==null)return r}catch{}return null},Re=e=>{if(!(e instanceof Element))return null;const n=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of n){const o=z(r);if(o!==null)return o}const i=e.matches("a[href]")?e:e.querySelector("a[href]");return i instanceof HTMLAnchorElement?se(i.getAttribute("href")):null},Ue=e=>{if(!e||typeof e!="object")return null;const n=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of n){const o=He(r);if(o!==null)return o}const i=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of i){const o=Re(r);if(o!==null)return o}return null},pe=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const i=document.querySelector(n);if(i instanceof HTMLElement)return i}return null},je=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+$+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute($,"1"));const i=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(i&&i.parentNode instanceof HTMLElement){const r=i.parentNode;return(n.parentNode!==r||n.nextSibling!==i)&&r.insertBefore(n,i),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},fe=e=>{const n=_e();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const i=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const o of i){const a=e.querySelector(o);if(!(a instanceof HTMLElement))continue;const s=[a.getAttribute("data-fileid"),a.getAttribute("data-file-id"),a.getAttribute("data-node-id"),a.getAttribute("data-id"),a.dataset&&(a.dataset.fileid||a.dataset.fileId)];for(const l of s){const c=z(l);if(c!==null)return c}}const r=e.querySelectorAll("a[href]");for(const o of r){if(!(o instanceof HTMLAnchorElement))continue;const a=se(o.getAttribute("href"));if(a!==null)return a}return null},ze=e=>e==="synced"?"etherpad-icon-color":Oe()?"etherpad-icon-white":"etherpad-icon-black",Be=e=>e==="synced"?t(d,"Synchronized"):e==="syncing"?t(d,"Saving..."):e==="error"?t(d,"Sync failed"):e==="unavailable"?t(d,"Sync status unavailable"):e==="loading"?t(d,"Checking sync status..."):t(d,"Sync pending"),x=()=>{const e=document.querySelector("["+B+'="1"]');e&&e.parentNode&&e.parentNode.removeChild(e);const n=document.querySelector("["+$+'="1"]');n&&n.parentNode&&n.parentNode.removeChild(n),F()},$e=(e,n)=>{if(!(e instanceof HTMLElement))return;const i=String(n||"").trim();e.setAttribute("data-public-open-url",i);const r=e.querySelector("[data-epnc-sidebar-open-public-button]");if(!(r instanceof HTMLButtonElement))return;const o=i!=="";r.style.display=o?"inline-flex":"none",r.disabled=!o},E=(e,n,i,r,o=!0)=>{if(!(e instanceof HTMLElement))return;const a=e.querySelector("[data-epnc-sidebar-sync-icon]"),s=e.querySelector("[data-epnc-sidebar-sync-label]"),l=e.querySelector("[data-epnc-sidebar-sync-button]");if(e.setAttribute("data-file-id",String(i)),e.setAttribute("data-sync-state",n),e.setAttribute("data-sync-in-flight",r?"1":"0"),s instanceof HTMLElement&&(s.textContent=Be(n)),a instanceof HTMLElement){const c=ze(n),u=V(d,c);u!==""&&(a.style.backgroundImage=`url(${u})`)}l instanceof HTMLButtonElement&&(l.disabled=r||!o||i<=0,l.setAttribute("aria-busy",r?"true":"false"))},De=(e,n)=>{if(!(n instanceof HTMLElement))return null;let i=n.querySelector("["+B+'="1"]');if(!(i instanceof HTMLElement)){i=document.createElement("section"),i.className="epnc-sidebar-sync-panel",i.setAttribute(B,"1"),i.setAttribute("data-file-id",String(e));const r=document.createElement("div");r.className="epnc-sidebar-sync-panel__header";const o=document.createElement("span");o.className="epnc-sidebar-sync-panel__icon",o.setAttribute("data-epnc-sidebar-sync-icon","1"),o.setAttribute("aria-hidden","true");const a=document.createElement("span");a.className="epnc-sidebar-sync-panel__status",a.setAttribute("data-epnc-sidebar-sync-label","1"),r.appendChild(o),r.appendChild(a);const s=document.createElement("button");s.type="button",s.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",s.setAttribute("data-epnc-sidebar-sync-button","1"),s.textContent=t(d,"Pad in Datei speichern"),s.addEventListener("click",async c=>{c.preventDefault(),c.stopPropagation();const u=Number(i.getAttribute("data-file-id")||"");if(!(!Number.isFinite(u)||u<=0)){E(i,"syncing",u,!0,!0);try{await We(u,!0);const h=await W(u),g=D(h&&h.status);E(i,g,u,!1,!0)}catch{E(i,"error",u,!1,!0)}}});const l=document.createElement("button");l.type="button",l.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",l.setAttribute("data-epnc-sidebar-open-public-button","1"),l.textContent=t(d,"Pad in neuem Tab öffnen"),l.style.display="none",l.addEventListener("click",c=>{c.preventDefault(),c.stopPropagation();const u=String(i.getAttribute("data-public-open-url")||"").trim();u!==""&&window.open(u,"_blank","noopener,noreferrer")}),i.appendChild(r),i.appendChild(s),i.appendChild(l)}return i.parentNode!==n&&n.appendChild(i),i},D=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",W=async e=>{const n=b("/apps/"+d+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e))),i=await fetch(n,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}),r=await i.json().catch(()=>({}));if(!i.ok)throw new Error(r&&r.message||"Sync status check failed.");return r},We=async(e,n)=>{let i=b("/apps/"+d+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));i+="?force=1";const r=await fetch(i,{method:"POST",headers:{Accept:"application/json",requesttoken:G()},credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Pad sync failed.");return o},F=()=>{O!==null&&(window.clearTimeout(O),O=null)},Ke=(e,n)=>e==="synced"?Math.min(xe,n+Ae):e==="pending"?ee:te,me=(e,n)=>{if(F(),!(e instanceof HTMLElement)||!Number.isFinite(n)||n<=0)return;let i=ee;const r=a=>{O=window.setTimeout(()=>{o()},Math.max(1e3,a))},o=async()=>{if(!(e instanceof HTMLElement)||!document.body.contains(e)){F();return}const a=pe();if(!(a instanceof HTMLElement)||!ue(a)||!ue(e)){F();return}if(fe(a)!==n){F();return}if(document.visibilityState!=="visible"||e.getAttribute("data-sync-in-flight")==="1"){r(i);return}try{const s=await W(n),l=D(s&&s.status);E(e,l,n,!1,!0),i=Ke(l,i)}catch{E(e,"error",n,!1,!0),i=te}r(i)};r(i)},Ve=()=>{S!==null&&(window.clearTimeout(S),S=null)},Ge=async()=>{const e=++_;if(!w()||q()){x();return}const n=pe();if(!(n instanceof HTMLElement)){x();return}const i=fe(n);if(!i){x();return}const r=je(n);if(!(r instanceof HTMLElement)){x();return}let o;try{o=await U(i)}catch{x();return}if(e!==_)return;if(!o||o.is_pad!==!0){x();return}const a=De(i,r);if(!(a instanceof HTMLElement)){x();return}const s=o&&typeof o.public_open_url=="string"?o.public_open_url.trim():"";$e(a,s),E(a,"loading",i,!1,!0);try{const l=await W(i);if(e!==_)return;const c=D(l&&l.status);E(a,c,i,!1,!0),me(a,i)}catch{if(e!==_)return;E(a,"error",i,!1,!0),me(a,i)}},K=(e=0)=>{Ve(),S=window.setTimeout(()=>{S=null,Ge()},Math.max(0,e))},Je=e=>{try{const n=new URL(e,window.location.origin),i=decodeURIComponent(n.pathname||"").match(/\/p\/([^/]+)$/);if(i&&i[1]){const r=i[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},Ze=async(e,n)=>{const i=b("/apps/"+d+"/api/v1/pads/from-url"),r=new URLSearchParams;r.set("file",e),r.set("padUrl",n);const o=await fetch(i,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:G()},credentials:"same-origin",body:r.toString()}),a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Could not import public pad URL.");return a},Qe=async e=>{const n=b("/apps/"+d+"/api/v1/pads"),i=new URLSearchParams;i.set("file",e),i.set("accessMode","public");const r=await fetch(n,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:G()},credentials:"same-origin",body:i.toString()}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Could not create public pad.");return o},ye=e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(d,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",i.appendChild(r),i.appendChild(o),n.appendChild(i),document.body.appendChild(n),{overlay:n,dialog:i,closeButton:r}},Xe=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=ye(t(d,"Internal public pad")),o=document.createElement("label");o.textContent=t(d,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(d,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const s=document.createElement("p");s.style.color="var(--color-error, #c62828)",s.style.margin="0 0 12px 0",s.style.minHeight="20px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(d,"Create");const c=u=>{n.remove(),e(u)};r.addEventListener("click",()=>c(null)),n.addEventListener("click",u=>{u.target===n&&c(null)}),a.addEventListener("keydown",u=>{u.key==="Enter"&&(u.preventDefault(),l.click())}),l.addEventListener("click",()=>{const u=a.value.trim();if(u===""){s.textContent=t(d,"File name is required.");return}c(u)}),i.appendChild(o),i.appendChild(a),i.appendChild(s),i.appendChild(l),a.focus(),a.select()}),Ye=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=ye(t(d,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(d,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const s=document.createElement("label");s.textContent=t(d,"File name"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="text",l.value="Imported pad.pad",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const c=document.createElement("p");c.style.color="var(--color-error, #c62828)",c.style.margin="0 0 12px 0",c.style.minHeight="20px";const u=document.createElement("button");u.type="button",u.className="primary",u.textContent=t(d,"Create");const h=y=>{n.remove(),e(y)};r.addEventListener("click",()=>h(null)),n.addEventListener("click",y=>{y.target===n&&h(null)}),a.addEventListener("blur",()=>{const y=a.value.trim();y.startsWith("http")&&(l.value=Je(y))});const g=()=>{const y=a.value.trim(),be=l.value.trim();if(y===""){c.textContent=t(d,"External pad URL is required.");return}if(be===""){c.textContent=t(d,"File name is required.");return}h({padUrl:y,name:be})};a.addEventListener("keydown",y=>{y.key==="Enter"&&(y.preventDefault(),g())}),l.addEventListener("keydown",y=>{y.key==="Enter"&&(y.preventDefault(),g())}),u.addEventListener("click",g),i.appendChild(o),i.appendChild(a),i.appendChild(s),i.appendChild(l),i.appendChild(c),i.appendChild(u),a.focus(),a.select()}),et=async()=>{const e=await Xe();if(!e)return;const n=e.trim(),i=n.toLowerCase().endsWith(".pad")?n:n+".pad",r=A(j(),i);try{const o=await Qe(r),a=o&&typeof o.file=="string"?o.file:r,s=o&&Number.isFinite(Number(o.file_id))?Number(o.file_id):null;if(w()&&s!==null){window.location.assign(N(s,a));return}await P({path:a,fileId:s})}catch(o){const a=o instanceof Error?o.message:t(d,"Could not create public pad.");window.alert(a)}},tt=async()=>{const e=await Ye();if(!e)return;const n=e.padUrl.trim(),i=e.name.trim(),r=j(),o=i.toLowerCase().endsWith(".pad")?i:i+".pad",a=A(r,o);try{const s=await Ze(a,n),l=s&&typeof s.file=="string"?s.file:a,c=s&&Number.isFinite(Number(s.file_id))?Number(s.file_id):null;if(w()&&c!==null){window.location.assign(N(c,l));return}await P({path:l,fileId:c})}catch(s){const l=s instanceof Error?s.message:t(d,"Could not import public pad URL.");window.alert(l)}},nt=()=>new Promise(e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(d,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(d,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(d,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const s=document.createElement("div");s.style.display="grid",s.style.gap="8px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(d,"Internal public pad");const c=document.createElement("button");c.type="button",c.textContent=t(d,"External public pad (URL)");const u=V(d,"etherpad-icon-color");u!==""&&(c.style.backgroundImage=`url(${u})`),c.style.backgroundRepeat="no-repeat",c.style.backgroundPosition="12px center",c.style.backgroundSize="16px 16px",c.style.paddingLeft="34px";const h=g=>{n.remove(),e(g)};o.addEventListener("click",()=>h(null)),l.addEventListener("click",()=>h("internal")),c.addEventListener("click",()=>h("external")),n.addEventListener("click",g=>{g.target===n&&h(null)}),s.appendChild(l),s.appendChild(c),i.appendChild(o),i.appendChild(r),i.appendChild(a),i.appendChild(s),n.appendChild(i),document.body.appendChild(n)}),it=async()=>{const e=await nt();if(e==="internal"){await et();return}e==="external"&&await tt()},rt=(e,n)=>{const i=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!i||typeof i.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(i.goToRoute(null,{view:"files",fileid:String(e)},{dir:de(n),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=N(e,n);window.location.assign(r)},180),!0)},P=async e=>{const n=String(e.fileId??"")+"|"+String(e.path??""),i=Date.now();if(J===n&&i-Z<800)return;J=n,Z=i;const r=q(),o=r!==null&&r!=="",a=()=>{if(o){window.location.assign(R(r,e.path||""));return}if(w()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const u=e.path||"/";window.location.assign(N(Number(e.fileId),u));return}const c=w()?I():null;if((e.fileId===null||e.fileId===void 0)&&c){const u=e.path||"/";window.location.assign(N(c,u));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const u=e.path||"/";window.location.assign(N(Number(e.fileId),u));return}e.path&&window.location.assign(Se(e.path))};if(!k()){a();return}let s=e.path||"",l=e.fileId??null;if(!s&&e.fileId!==null&&e.fileId!==void 0)try{const c=await U(e.fileId);s=c&&typeof c.path=="string"?c.path:"",l=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):l}catch{s=""}if(!s){a();return}if((!l||!Number.isFinite(Number(l)))&&s&&!o)try{const c=await Pe(s);l=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):l}catch{}if((!l||!Number.isFinite(Number(l)))&&w()){const c=I();c&&(l=c)}if(w()){if(l&&Number.isFinite(Number(l))&&rt(Number(l),s))return;a();return}try{window.OCA.Viewer.open({path:s})}catch{a()}},ot=(e,n)=>{if(!(e instanceof HTMLAnchorElement)||!n)return null;const i=e.getAttribute("href")||"",r=ke(i);if(r&&r.token===n&&M(r.path))return r.path;const o=vt(i);return M(o)?o:null},at=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const e=async n=>{if(!n||n.defaultPrevented)return;const i=q();if(!i||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const r=n.target.closest("a[href]");if(!(r instanceof HTMLAnchorElement))return;const o=ot(r,i);if(o){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),k()){await P({path:o,fileId:null});return}window.location.assign(R(i,o))}};document.addEventListener("click",e,!0)},st=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(ge,p,ht(),V(d,"etherpad-icon-color"),(e,n)=>{const i=n&&n.dir||j(),r=A(i,e),o=Ue(n);P({path:r,fileId:o})},t(d,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(ge,p))},lt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const i of e)if(!(!i||typeof i!="object")){if(typeof i.addNewFileMenuEntry=="function")return{register:i.addNewFileMenuEntry.bind(i),unregister:typeof i.removeNewFileMenuEntry=="function"?i.removeNewFileMenuEntry.bind(i):null};if(typeof i.getNewFileMenu=="function"){let r=null;try{r=i.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const n=window._nc_newfilemenu;return n&&typeof n=="object"&&typeof n.registerEntry=="function"?{register:n.registerEntry.bind(n),unregister:typeof n.unregisterEntry=="function"?n.unregisterEntry.bind(n):null}:null},ct=async()=>{const e=lt();if(!e)return!1;const n=ce();try{return e.register(n),H=!0,!0}catch(i){return T(i)?(H=!0,!0):!1}},dt=()=>{if(C)return!0;const e=Fe(),n=()=>{const i=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const r of i)if(!(!r||typeof r!="object")){if(typeof r.addMenuEntry=="function")try{return r.addMenuEntry(e),!0}catch(o){if(T(o))return!0}if(typeof r.registerEntry=="function")try{return r.registerEntry(ce()),!0}catch(o){if(T(o))return!0}}return!1};try{return n()?(C=!0,!0):(!X&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(i){if(!(!i||typeof i.addMenuEntry!="function"))try{i.addMenuEntry(e),C=!0}catch(r){T(r)&&(C=!0)}}}),X=!0),C)}catch(i){return T(i)?(C=!0,!0):!1}},ut=()=>{if(!w()||H||C)return;const e=++Y,n=async i=>{if(!(e!==Y||H||C||await ct()||dt())&&i<Ne){window.setTimeout(()=>{n(i+1)},Le);return}};n(0)},pt=()=>{if(v!==null||!w())return;const e=new MutationObserver(n=>{for(const i of n)for(const r of i.addedNodes)if(r instanceof Element&&(r.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||r.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){K(60);return}});e.observe(document.body,{childList:!0,subtree:!0}),v=e},ft=async()=>{if(k())return;const e=I();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("epNoRedirect")==="1"||n.get("details")==="1"||n.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+d+"/"))try{const i=await U(e);i&&i.is_pad&&i.viewer_url&&window.location.assign(i.viewer_url)}catch{}},mt=async()=>{if(k()||!w())return;const e=I();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("openfile")==="true"||n.get("opendetails")==="true"||n.get("details")==="1"))try{const i=await U(e);if(!i||!i.is_pad)return;const r=n.get("dir")||"/",o=b("/apps/files/files")+"?dir="+encodeURIComponent(r);window.location.href!==o&&window.location.assign(o)}catch{}},yt=()=>{const e=q();if(!e)return;const n=new URLSearchParams(window.location.search||""),i=n.get("files")||"";if(k()){if(!M(i))return;const a=n.get("path")||"/",s=A(a,i);P({path:s,fileId:null});return}if(!M(i)){if((window.location.pathname||"").includes("/apps/"+d+"/public/"))return;Ee(e)&&window.location.assign(R(e,""));return}const r=n.get("path")||"/",o=A(r,i);window.location.assign(R(e,o))},we=()=>{const e=(window.location.pathname||"")+"?"+(window.location.search||"");e!==Q&&(Q=e,w()&&ut(),xt(),mt(),ft(),yt(),K(0))},wt=()=>{const e="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[e]===!0)return;window.OCA&&(window.OCA[e]=!0);const n=()=>window.setTimeout(we,0);window.addEventListener("popstate",n),window.addEventListener("hashchange",n);const i=r=>{const o=window.history&&window.history[r];typeof o=="function"&&(window.history[r]=function(...a){const s=o.apply(this,a);return n(),s})};i("pushState"),i("replaceState")},he=()=>{m||(m=!0,wt(),we(),st(),at(),pt(),K(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",he,{once:!0}):he()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
